Add displayconf and screenlayout
The displayconf script enables the user to quickly change between
diffrent screenlayouts. It opens up a dmenu where the diffrent layouts
are listed. Those layouts are executable scripts which will change the
screenlayout. The layout scripts are gathered from the $LAYOUTDIR path.
A few sample scripts are added to the /screenlayout directory.